+++ Assets/Scripts/Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 64e806e80901441629bdd800a10592f5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
