--- Assets/Scripts/RoomDesignerEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 28ca9e159b737411da3e731aa5b14ae2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
