--- Assets/Scripts/StageManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3026e9c3efe854d0490feb26a54dfe34
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
