+++ Assets/Scripts/TurnTableSet.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: edb2dd0302ace43f2b35dc45effc9504
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
