+++ Assets/Sprites/GameTurn.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 08bf42186e51041f9b021fe4218b5e46
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
